move private function to internal pkg for sharing
